multi added

+++ app/src/main/java/com/github/displace/sdp2022/GameVersusViewActivity.kt
@@ -1,10 +1,12 @@
 package com.github.displace.sdp2022
 
 import android.content.Intent
+import android.os.Build
 import android.os.Bundle
 import android.view.View
 import android.widget.EditText
 import android.widget.TextView
+import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatActivity
 import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.recyclerview.widget.LinearLayoutManager
@@ -15,6 +17,8 @@ import com.github.displace.sdp2022.gameVersus.Chat
 import com.github.displace.sdp2022.gameVersus.ClientServerLink
 import com.github.displace.sdp2022.gameVersus.GameVersusViewModel
 import com.github.displace.sdp2022.map.*
+import com.github.displace.sdp2022.profile.FriendRequest
+import com.github.displace.sdp2022.users.PartialUser
 import com.github.displace.sdp2022.util.DateTimeUtil
 import com.github.displace.sdp2022.util.PreferencesUtil
 import com.github.displace.sdp2022.util.gps.GPSPositionManager
@@ -26,6 +30,8 @@ import com.google.firebase.database.*
 import org.osmdroid.util.GeoPoint
 import org.osmdroid.views.MapView
 import kotlin.collections.ArrayList
+import com.github.displace.sdp2022.util.math.Constants
+import java.io.Serializable
 
 
 const val EXTRA_STATS = "com.github.displace.sdp2022.GAMESTAT"
@@ -48,11 +54,14 @@ class GameVersusViewActivity : AppCompatActivity() {
     private lateinit var gpsPositionUpdater: GPSPositionUpdater
     private lateinit var gpsPositionManager: GPSPositionManager
     private lateinit var other: Map<String, Any>
+    private var others = listOf<List<String>>()
     private lateinit var pinpointsDBHandler: PinpointsDBHandler
     private lateinit var pinpointsManager: PinpointsManager
     private lateinit var otherPlayerPinpoints: PinpointsManager.PinpointsRef
     private lateinit var conditionalGoalPlacer: ConditionalGoalPlacer
     private lateinit var clientServerLink : ClientServerLink
+
+    private var nbPlayer = 1L
 
     //CHAT
     private lateinit var chat : Chat
@@ -70,44 +79,47 @@ class GameVersusViewActivity : AppCompatActivity() {
     // 3 possibility : win => guess == position of the goal, continue => guess != position and lost => you missed the max number of time and lost
     private val guessListener = GeoPointListener { geoPoint ->
         run {
-            pinpointsManager.putMarker(geoPoint)
-            val res = game.handleEvent(
-                GameEvent.OnPointSelected(
-                    intent.getStringExtra("uid")!!,
-                    Point(geoPoint.latitude, geoPoint.longitude)
+            if (CoordinatesUtil.distance(game.getPos(), CoordinatesUtil.coordinates(geoPoint)) <= Constants.CLICKABLE_AREA_RADIUS)
+            {
+                pinpointsManager.putMarker(geoPoint)
+                val res = game.handleEvent(
+                    GameEvent.OnPointSelected(
+                        intent.getStringExtra("uid")!!,
+                        Point(geoPoint.latitude, geoPoint.longitude)
+                    )
                 )
-            )
-            if (res == GameVersusViewModel.WIN) {
-                gpsPositionManager.listenersManager.clearAllCalls()
-                clientServerLink.listenerManager.clearAllCalls()
-                findViewById<TextView>(R.id.TryText).apply { text = "win" }
-                extras.putBoolean(EXTRA_RESULT, true)
-                extras.putInt(EXTRA_SCORE_P1, 1)
-                extras.putInt(EXTRA_SCORE_P2, 0)
-                statsList.clear()
-                statsList.add(DateTimeUtil.currentTime())
-                showGameSummaryActivity()
-            } else if (res == GameVersusViewModel.CONTINUE) {
-                findViewById<TextView>(R.id.TryText).apply {
-                    text = "wrong guess, remaining tries : ${4 - game.getNbEssai()}"
+                if (res == GameVersusViewModel.WIN) {
+                    gpsPositionManager.listenersManager.clearAllCalls()
+                    clientServerLink.listenerManager.clearAllCalls()
+                    findViewById<TextView>(R.id.TryText).apply { text = "win" }
+                    extras.putBoolean(EXTRA_RESULT, true)
+                    extras.putInt(EXTRA_SCORE_P1, 1)
+                    extras.putInt(EXTRA_SCORE_P2, 0)
+                    statsList.clear()
+                    statsList.add(DateTimeUtil.currentTime())
+                    showGameSummaryActivity()
+                } else if (res == GameVersusViewModel.CONTINUE) {
+                    findViewById<TextView>(R.id.TryText).apply {
+                        text = "wrong guess, remaining tries : ${4 - game.getNbEssai()}"
+                    }
+                    pinpointsDBHandler.updateDBPinpoints(
+                        intent.getStringExtra("uid")!!,
+                        pinpointsManager.playerPinPointsRef
+                    )
+                } else if (res == GameVersusViewModel.LOSE) {
+                    gpsPositionManager.listenersManager.clearAllCalls()
+                    clientServerLink.listenerManager.clearAllCalls()
+                    findViewById<TextView>(R.id.TryText).apply {
+                        text =
+                            "status : end of game"
+                    }
+                    extras.putBoolean(EXTRA_RESULT, false)
+                    extras.putInt(EXTRA_SCORE_P1, 0)
+                    extras.putInt(EXTRA_SCORE_P2, 1)
+                    statsList.clear()
+                    statsList.add(DateTimeUtil.currentTime())
+                    showGameSummaryActivity()
                 }
-                pinpointsDBHandler.updateDBPinpoints(
-                    intent.getStringExtra("uid")!!,
-                    pinpointsManager.playerPinPointsRef
-                )
-            } else if (res == GameVersusViewModel.LOSE) {
-                gpsPositionManager.listenersManager.clearAllCalls()
-                clientServerLink.listenerManager.clearAllCalls()
-                findViewById<TextView>(R.id.TryText).apply {
-                    text =
-                        "status : end of game"
-                }
-                extras.putBoolean(EXTRA_RESULT, false)
-                extras.putInt(EXTRA_SCORE_P1, 0)
-                extras.putInt(EXTRA_SCORE_P2, 1)
-                statsList.clear()
-                statsList.add(DateTimeUtil.currentTime())
-                showGameSummaryActivity()
             }
         }
     }
@@ -116,7 +128,10 @@ class GameVersusViewActivity : AppCompatActivity() {
     private val endListener = object : ValueEventListener {
         override fun onDataChange(dataSnapshot: DataSnapshot) {
             val x = dataSnapshot.value
-            if (x == GameVersusViewModel.LOSE || x == GameVersusViewModel.WIN) {
+            if(x == GameVersusViewModel.LOSE){
+                nbPlayer -= 1
+            }
+            if ((x == GameVersusViewModel.LOSE && nbPlayer < 2) || x == GameVersusViewModel.WIN) {
                 db.delete("GameInstance", "Game" + intent.getStringExtra("gid")!!)
                 gpsPositionManager.listenersManager.clearAllCalls()
                 clientServerLink.listenerManager.clearAllCalls()
@@ -155,6 +170,7 @@ class GameVersusViewActivity : AppCompatActivity() {
         ) as RealTimeDatabase
         clientServerLink = ClientServerLink(db)
         game = GameVersusViewModel(clientServerLink)
+        nbPlayer = intent.getLongExtra("nbPlayer",1)
 
         //initialise all the viewer and manager.
         mapView = findViewById(R.id.map)
@@ -238,13 +254,10 @@ class GameVersusViewActivity : AppCompatActivity() {
     //show the game sumarry by launching a new activity
     private fun showGameSummaryActivity() {
         val intent = Intent(this, GameSummaryActivity::class.java)
-        val otherPlayerId =
-            other.toList()[0].first.removePrefix("id:")
-        extras.putString( "OPPONENT_ID" ,otherPlayerId )
-
         extras.putStringArrayList(EXTRA_STATS, statsList)
         extras.putString(EXTRA_MODE, "Versus")
         intent.putExtras(extras)
+        intent.putExtra("others",others as Serializable)
         startActivity(intent)
     }
 
@@ -256,25 +269,34 @@ class GameVersusViewActivity : AppCompatActivity() {
             }"
         }
 
-        val otherPlayerId =
-            other.toList()[0].first.removePrefix("id:") // currently supporting only 2 players
-        db.addList(
-            "GameInstance/Game${intent.getStringExtra("gid")!!}/id:${otherPlayerId}",
-            "finish",
-            endListener
-        )
-        pinpointsDBHandler.enableAutoupdateLocalPinpoints(
-            otherPlayerId,
-            otherPlayerPinpoints
-        )
-        game.handleEvent(
-            GameEvent.OnStart(
-                DEFAULT_GOAL,
-                intent.getStringExtra("uid")!!,
-                intent.getStringExtra("gid")!!,
-                otherPlayerId
+        other.toList().forEach { x ->
+            val otherPlayerId = x.first.removePrefix("id:") // currently supporting only 2 players
+            db.addList(
+                "GameInstance/Game${intent.getStringExtra("gid")!!}/id:${otherPlayerId}",
+                "finish",
+                endListener
             )
-        )
+
+            db.referenceGet("CompleteUsers/${otherPlayerId}/CompleteUser/partialUser", "username").addOnSuccessListener { snapshot ->
+                val name = snapshot.value as String
+                val list = listOf(listOf(name,otherPlayerId))
+                others = others.plus(list)
+            }
+
+            pinpointsDBHandler.enableAutoupdateLocalPinpoints(
+                otherPlayerId,
+                otherPlayerPinpoints
+            )
+
+            game.handleEvent(
+                GameEvent.OnStart(
+                    DEFAULT_GOAL,
+                    intent.getStringExtra("uid")!!,
+                    intent.getStringExtra("gid")!!,
+                    otherPlayerId
+                )
+            )
+        }
     }
 
     companion object {
